fix(engine): drop-in parity for network_mode, selinux binds, link warnings (#491)
## Summary

Drop-in parity correctness fixes so `docker-compose.yml` behaves the
same under podup as under docker-compose v2. Four independent fixes, one
logical commit each.

### 1. `network_mode: service:<name>` on a scaled target (HIGH)
The target was resolved to the base name `{project}-{svc}`. When that
service is scaled (`scale:` or `deploy.replicas` > 1) the base-named
container does not exist — its replicas are `{project}-{svc}-1..N` — so
Podman rejected the netns reference. Now resolves to replica `-1` for
scaled targets (matching docker-compose). Explicit `container_name` is
still honoured verbatim; single-replica behaviour is unchanged.

### 2. `bind.selinux: shared/private` not translated (MED)
The literal Compose words were forwarded as Podman mount options, which
Podman does not recognise. Now mapped `shared`->`z`, `private`->`Z`; any
other value passes through verbatim.

### 3. `external_links` missing rootless caveat (MED)
`links` emitted a rootless-networking caveat warning but
`external_links` did not, despite the same no-effect behaviour under
rootless Podman. Added an analogous advisory.

### 4. `gw_priority` engine-time log invisible (LOW)
The engine-time notice was a `tracing::debug!` (invisible at default
level) while the user-facing notice is already emitted once at parse
time by the compose diagnostics. Removed the engine-time line to avoid a
silent duplicate; the parse-time diagnostic still fires.

## Tests

Unit tests added for the pure parts: scaled-target replica-name
resolution (single/scale/deploy.replicas/container_name/unknown-target)
and the selinux value mapping (shared->z, private->Z, passthrough). No
podman integration run here.

`cargo fmt --all -- --check`, `cargo clippy --all-targets --all-features
-- -D warnings`, `cargo build --all-features`, and `cargo test --lib
--all-features` (660 passed) all pass.

Closes #479

---------
